#3 split szRecord into it's own file with tests
